refactor(challenge): split ChallengePanel into challenge/ subcomponents (#102 B)
