+++ src/components/challenge/DrillPanel.tsx
@@ -1,0 +1,239 @@
+import { ArrowRight, Eye, GraduationCap, RotateCcw } from "lucide-react";
+import { copy, type Language } from "../../i18n";
+import type { PartOfSpeech } from "../../domain/types";
+import { DarumaSpot, PaperNoteSpot, TeaCupSpot } from "../../illustrations";
+import { ExamPrompt } from "../ExamPrompt";
+import { FeedbackPanel } from "../FeedbackPanel";
+import { SpeakButton } from "../SpeakButton";
+import type { Feedback } from "../types";
+import type { PracticeSession } from "../../hooks/usePracticeSession";
+
+function choiceOptionClass(choice: string, selectedChoice: string | null, feedback: Feedback): string {
+  const classes = ["choice-option"];
+
+  if (selectedChoice === choice) {
+    classes.push("chosen");
+
+    if (feedback?.status === "correct") {
+      classes.push("correct");
+    }
+
+    if (feedback?.status === "incorrect") {
+      classes.push("incorrect");
+    }
+  }
+
+  return classes.join(" ");
+}
+
+function partOfSpeechLabel(partOfSpeech: PartOfSpeech, language: Language): string {
+  return copy[language].partOfSpeech[partOfSpeech];
+}
+
+// The centre column of the challenge workspace: the active drill. Renders
+// the current question's prompt + choice grid + action row + post-answer
+// feedback, or one of the three terminal states (session exhausted /
+// review queue empty / nothing to practise). Pure presentation over the
+// practice session's state + handlers; extracted from ChallengePanel with
+// no behavioural change. `onExit` returns to the home dashboard from the
+// completion / empty screens.
+export function DrillPanel({
+  language,
+  questionIndex,
+  selectedChoice,
+  feedback,
+  attempts,
+  practiceMode,
+  currentQuestion,
+  reviewEmpty,
+  sessionExhausted,
+  choiceOptions,
+  correctCount,
+  nextButtonRef,
+  setPracticeMode,
+  setPracticeFilter,
+  handleChoiceSubmit,
+  nextQuestion,
+  resetSession,
+  revealAnswer,
+  handleDrillKeyDown,
+  onExit
+}: Pick<
+  PracticeSession,
+  | "questionIndex"
+  | "selectedChoice"
+  | "feedback"
+  | "attempts"
+  | "practiceMode"
+  | "currentQuestion"
+  | "reviewEmpty"
+  | "sessionExhausted"
+  | "choiceOptions"
+  | "correctCount"
+  | "nextButtonRef"
+  | "setPracticeMode"
+  | "setPracticeFilter"
+  | "handleChoiceSubmit"
+  | "nextQuestion"
+  | "resetSession"
+  | "revealAnswer"
+  | "handleDrillKeyDown"
+> & { language: Language; onExit: () => void }) {
+  const t = copy[language];
+
+  // Container-level answer state for embedded AI / browser automation:
+  // collapse feedback into one result string so .drill-panel exposes the
+  // whole state (which question, what was picked, the outcome) in one place
+  // without having to scan the option buttons. "unanswered" until answered.
+  const drillResult = !feedback
+    ? "unanswered"
+    : feedback.status === "correct"
+      ? "correct"
+      : feedback.status === "revealed"
+        ? "revealed"
+        : "wrong";
+
+  return (
+    <section
+      className="drill-panel"
+      aria-label={t.currentQuestion}
+      onKeyDown={handleDrillKeyDown}
+      data-question-id={currentQuestion?.id}
+      data-question-type={currentQuestion?.promptLabel ?? currentQuestion?.targetForm}
+      data-selected={selectedChoice ?? undefined}
+      data-result={drillResult}
+      data-expected-answer={feedback ? currentQuestion?.expectedAnswers.join(" / ") : undefined}
+    >
+      {currentQuestion ? (
+        <>
+          <div className="prompt-header">
+            <span>{t.questionNumber(questionIndex + 1)}</span>
+            <strong>{currentQuestion.promptLabel ?? t.targetForms[currentQuestion.targetForm]}</strong>
+          </div>
+
+          <div className="word-block">
+            {currentQuestion.promptText ? (
+              <ExamPrompt question={currentQuestion} language={language} />
+            ) : (
+              <>
+                <p className="word-kind">
+                  <GraduationCap aria-hidden="true" />
+                  {partOfSpeechLabel(currentQuestion.vocabulary.partOfSpeech, language)}
+                </p>
+                {currentQuestion.targetForm === "reading" ? null : (
+                  <p className="reading">{currentQuestion.vocabulary.reading}</p>
+                )}
+                <p className="surface">
+                  {currentQuestion.vocabulary.surface}
+                  <SpeakButton
+                    text={currentQuestion.vocabulary.surface}
+                    language={language}
+                  />
+                </p>
+                {currentQuestion.targetForm === "meaning" ? null : (
+                  <p className="meaning">{currentQuestion.vocabulary.meaningZh}</p>
+                )}
+              </>
+            )}
+          </div>
+
+          <div className="choice-grid" aria-label={t.answerOptions}>
+            {choiceOptions.map((choice) => {
+              // Expose selection + result as DOM data attributes for AI /
+              // browser-automation testability. Derived purely from existing
+              // state -- no change to click handling or to choiceOptionClass
+              // (the visual styling). Multi-answer aware via expectedAnswers.
+              const isSelected = selectedChoice === choice;
+              let dataResult: "correct" | "wrong" | "target" | undefined;
+              if (feedback) {
+                if (isSelected) {
+                  dataResult = feedback.status === "correct" ? "correct" : "wrong";
+                } else if (
+                  feedback.status !== "correct" &&
+                  currentQuestion.expectedAnswers.includes(choice)
+                ) {
+                  // Got it wrong (or revealed) -> flag the correct answer button.
+                  dataResult = "target";
+                }
+              }
+              return (
+                <button
+                  key={choice}
+                  type="button"
+                  className={choiceOptionClass(choice, selectedChoice, feedback)}
+                  disabled={Boolean(feedback)}
+                  onClick={() => handleChoiceSubmit(choice)}
+                  data-selected={isSelected ? "true" : undefined}
+                  data-result={dataResult}
+                >
+                  {choice}
+                </button>
+              );
+            })}
+          </div>
+
+          <div className="action-row">
+            <button className="ghost-button" type="button" onClick={revealAnswer} disabled={Boolean(feedback)}>
+              <Eye aria-hidden="true" />
+              {t.revealAnswer}
+            </button>
+            <button className="next-button" type="button" ref={nextButtonRef} onClick={nextQuestion}>
+              <ArrowRight aria-hidden="true" />
+              {t.nextQuestion}
+            </button>
+          </div>
+
+          {feedback ? (
+            <FeedbackPanel feedback={feedback} language={language} options={choiceOptions} />
+          ) : null}
+        </>
+      ) : sessionExhausted ? (
+        <div className="empty-state review-done">
+          <DarumaSpot />
+          <h2>{practiceMode === "daily" ? t.dailyDoneTitle : t.reviewDoneTitle}</h2>
+          <p>
+            {practiceMode === "daily"
+              ? t.dailyDoneBody(correctCount, attempts.length - correctCount)
+              : t.reviewDoneBody(correctCount, attempts.length - correctCount)}
+          </p>
+          <div className="review-done-actions">
+            <button className="next-button" type="button" onClick={resetSession}>
+              <RotateCcw aria-hidden="true" />
+              {practiceMode === "daily" ? t.dailyDoneAgain : t.reviewDoneAgain}
+            </button>
+            <button className="ghost-button" type="button" onClick={onExit}>
+              {practiceMode === "daily" ? t.dailyDoneExit : t.reviewDoneExit}
+            </button>
+          </div>
+        </div>
+      ) : reviewEmpty ? (
+        <div className="empty-state review-done">
+          <TeaCupSpot />
+          <p>{t.reviewEmptyState}</p>
+          <div className="review-done-actions">
+            <button
+              className="next-button"
+              type="button"
+              onClick={() => {
+                setPracticeMode("exam");
+                setPracticeFilter({});
+                resetSession();
+              }}
+            >
+              <ArrowRight aria-hidden="true" />
+              {t.reviewEmptyCta}
+            </button>
+            <button className="ghost-button" type="button" onClick={onExit}>
+              {t.reviewDoneExit}
+            </button>
+          </div>
+        </div>
+      ) : (
+        <div className="empty-state empty-state-illustrated">
+          <PaperNoteSpot />
+          <p>{t.emptyState}</p>
+        </div>
+      )}
+    </section>
+  );
+}
